+++ core/utils/cli_options.ts
@@ -8,7 +8,7 @@ export function getOptionsFromCli(
   const cli = parseArgs(Deno.args, {
     string: ["src", "dest", "location", "port"],
     boolean: ["serve", "open"],
-    alias: { dev: "d", serve: "s", port: "p", open: "o" },
+    alias: { serve: "s", port: "p", open: "o" },
     ["--"]: true,
   });
 
@@ -22,40 +22,48 @@ export function getOptionsFromCli(
 
   const serveMode = cli.serve || cli._[0] === "cms";
 
-  if (cli.port) {
-    (options.server ||= {}).port = parseInt(cli.port);
-  } else if (serveMode) {
-    (options.server ||= {}).port = 3000;
-  }
-
+  // Detect location and port
+  let port: number;
   let location: URL;
 
-  if (cli.location) {
-    location = new URL(cli.location);
-  } else if (options.location && !serveMode) {
-    location = options.location as URL;
+  if (serveMode) {
+    location = cli.location
+      ? new URL(cli.location)
+      : new URL("http://localhost");
+
+    if (cli.port) {
+      port = parseInt(cli.port);
+      location.port = port.toString();
+    } else if (location.port) {
+      port = parseInt(location.port);
+    } else if (options.server?.port) {
+      port = options.server.port;
+      location.port = port.toString();
+    } else {
+      port = location.protocol === "https:" ? 443 : 3000;
+      location.port = port.toString();
+    }
   } else {
-    location = new URL("http://localhost");
+    location = cli.location
+      ? new URL(cli.location)
+      : (options.location as URL | undefined) || new URL("http://localhost");
+
+    if (cli.port) {
+      port = parseInt(cli.port);
+      location.port = port.toString();
+    } else if (location.port) {
+      port = parseInt(location.port);
+    } else {
+      port = location.protocol === "https:" ? 443 : 80;
+    }
   }
 
-  let port: number;
-
-  if (cli.port) {
-    port = parseInt(cli.port);
-  } else if (location.port) {
-    port = parseInt(location.port);
-  } else if (options.server?.port) {
-    port = options.server.port;
-  } else {
-    port = cli.serve ? 3000 : location.protocol === "https:" ? 443 : 80;
-  }
-
-  (options.server ||= {}).port = port;
-  location.port = port.toString();
   options.location = location;
+  options.server ||= {};
+  options.server.port = port;
 
   if (cli.open) {
-    (options.server ||= {}).open = cli.open;
+    options.server.open = cli.open;
   }
 
   return options;