+++ core/utils/net.ts
@@ -1,9 +1,14 @@
 export function localIp(): string | undefined {
-  for (const info of Deno.networkInterfaces()) {
-    if (info.family !== "IPv4" || info.address.startsWith("127.")) {
-      continue;
-    }
+  // Try/catch for https://github.com/denoland/deno/issues/25420
+  try {
+    for (const info of Deno.networkInterfaces()) {
+      if (info.family !== "IPv4" || info.address.startsWith("127.")) {
+        continue;
+      }
 
-    return info.address;
+      return info.address;
+    }
+  } catch {
+    return undefined;
   }
 }