+++ CHANGELOG.md
@@ -8,17 +8,27 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [Unreleased]
 ### Added
 - New plugin: `json_ld` for generating JSON-LD tags in the output [#453]
-
-## [2.4.3] - Unreleased
-### Added
 - New plugin: `purgecss` to remove unused CSS. [#693]
+
+## [2.4.3] - 2024-12-11
+### Added
+- New option `finalHandler` to `Server` class.
+- `metas` plugin: Support for light and dark theme colors using an array.
+  Example:
+  ```yml
+  color: ["white", "black"]
+  ```
 
 ### Fixed
 - Ensure `LUME_LIVE_RELOAD` env variable is available in the _config file.
 - `on_demand` middleware types.
 - `on_demand` plugin doesn't filter scoped pages.
 - Added test for `check_urls` plugin.
-- Updates dependencies: `std`, `sass`, `liquidjs`, `tailwindcss`, `postcss`, `decap-cms`, `unocss`, `magic-string` and some icons.
+- Added tests for middlewares.
+- Types errors in Deno 2.1.
+- `og_images` plugin: make satori options type partial.
+- Fixed `--port` configuration on serve mode [#704].
+- Updates dependencies: `std`, `sass`, `liquidjs`, `tailwindcss`, `postcss`, `lightningcss`, `preact`, `decap-cms`, `unocss`, `magic-string`, `vento`, `satori`, `terser`, `markdown-it-attrs`, `katex`, `xml` and some icons.
 
 ## [2.4.2] - 2024-11-10
 ### Fixed
@@ -620,8 +630,10 @@ Go to the `v1` branch to see the changelog of Lume 1.
 [#686]: https://github.com/lumeland/lume/issues/686
 [#689]: https://github.com/lumeland/lume/issues/689
 [#693]: https://github.com/lumeland/lume/issues/693
+[#704]: https://github.com/lumeland/lume/issues/704
 
-[2.4.3]: https://github.com/lumeland/lume/compare/v2.4.2...HEAD
+[Unreleased]: https://github.com/lumeland/lume/compare/v2.4.3...HEAD
+[2.4.3]: https://github.com/lumeland/lume/compare/v2.4.2...v2.4.3
 [2.4.2]: https://github.com/lumeland/lume/compare/v2.4.1...v2.4.2
 [2.4.1]: https://github.com/lumeland/lume/compare/v2.4.0...v2.4.1
 [2.4.0]: https://github.com/lumeland/lume/compare/v2.3.3...v2.4.0